# docs/swagger.yaml
definitions:
...
  json.LojaResponse:
    properties:
      categoria:
        type: string
      cnpj:
        type: string
      distancia:
        description: Distância em km
        type: number
      id:
        type: integer
      id_categoria:
        type: integer
      imagem:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      nome:
        type: string
    type: object
...
  json.ProdutoRequest:
    properties:
      descricao:
        type: string
      estoque:
        minimum: 0
        type: integer
      id_loja:
        type: integer
      imagem:
        type: string
      nome:
        type: string
      preco:
        minimum: 0
        type: number
    required:
    - id_loja
    - nome
    - preco
    type: object
  json.ProdutoResponse:
    properties:
      ativo:
        type: boolean
      data_cadastro:
        type: string
      descricao:
        type: string
      estoque:
        type: integer
      id:
        type: integer
      id_loja:
        type: integer
      imagem:
        type: string
      loja:
        $ref: '#/definitions/json.LojaResponse'
      nome:
        type: string
      preco:
        type: number
    type: object
  json.ProdutosResponse:
    properties:
      produtos:
        items:
          $ref: '#/definitions/json.ProdutoResponse'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /lojas/{id}/produtos:
    get:
      consumes:
      - application/json
      description: Retorna todos os produtos ativos de uma loja específica
      parameters:
      - description: ID da loja
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.ProdutosResponse'
        "400":
          description: ID de loja inválido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista produtos de uma loja
      tags:
      - Produtos
...
  /produtos:
    get:
      consumes:
      - application/json
      description: Retorna todos os produtos ativos do sistema
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.ProdutosResponse'
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Lista todos os produtos
      tags:
      - Produtos
    post:
      consumes:
      - application/json
      description: Cria um novo produto no sistema
      parameters:
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/json.ProdutoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/json.ProdutoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Cria um novo produto
      tags:
      - Produtos
  /produtos/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza soft delete de um produto, marcando-o como excluído
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Produto removido com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Remove produto (soft delete)
      tags:
      - Produtos
    get:
      consumes:
      - application/json
      description: Retorna um produto específico pelo ID
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.ProdutoResponse'
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Busca produto por ID
      tags:
      - Produtos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto existente
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/json.ProdutoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/json.ProdutoResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza produto
      tags:
      - Produtos
...
  /produtos/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restaura um produto que foi soft deleted
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Produto restaurado com sucesso
          schema:
            additionalProperties: true
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties: true
            type: object
      summary: Restaura produto
      tags:
      - Produtos
...
swagger: "2.0"